refactor: refactored component types
