+++ src/types/component-types/index.ts
@@ -5,3 +5,5 @@ export type { authInputType, errorType } from "./auth.types";
 export type { asideNavBarType } from "./asideNavBar.type";
 export type { resultCardType } from "./resultCard.type";
 export type { loaderType } from "./loader.type";
+export type { KnowledgeBoardType } from "./knowledgeBoard.type";
+export type { profileCardType } from "./profileCard.type";
