+++ src/types/component-types/knowledgeBoard.type.ts
@@ -1,0 +1,6 @@
+type KnowledgeBoardType = {
+  quizzesAttempted: number;
+  totalScore: number;
+};
+
+export type { KnowledgeBoardType };
