+++ src/types/component-types/profileCard.type.ts
@@ -1,0 +1,7 @@
+type profileCardType = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  signOut: () => void;
+};
+export type { profileCardType };
